CXO-7162 - Typed properties having integerSpec/longSpec are shown as text fields in UI

+++ copl/com.elster.jupiter.properties.rest/src/main/java/com/elster/jupiter/properties/rest/SimplePropertyType.java
@@ -66,7 +66,7 @@ public enum SimplePropertyType implements PropertyType {
             if (simplePropertyType.matches(valueFactory)) {
                 return simplePropertyType;
             }
-            if(valueFactory.getValueType().equals(Integer.class) || valueFactory.getValueType().equals(Integer.class)){
+            if(valueFactory.getValueType().equals(Integer.class) || valueFactory.getValueType().equals(Long.class)){
                 return LONG;
             }
         }
